feat: add new pages and bootstrap
