+++ src/views/pages/Error404.ts
